@@ -16,19 +16,20 @@ const dialTimeout = 10 * time.Second
 // TLSPool is a connections pool for the DNS-over-TLS Upstream.
 //
 // Example:
-//  pool := TLSPool{Address: "tls://1.1.1.1:853"}
-//  netConn, err := pool.Get()
-//  if err != nil {panic(err)}
-//  c := dns.Conn{Conn: netConn}
-//  q := dns.Msg{}
-//  q.SetQuestion("google.com.", dns.TypeA)
-//  log.Println(q)
-//  err = c.WriteMsg(&q)
-//  if err != nil {panic(err)}
-//  r, err := c.ReadMsg()
-//  if err != nil {panic(err)}
-//  log.Println(r)
-//  pool.Put(c.Conn)
+//
+//	pool := TLSPool{Address: "tls://1.1.1.1:853"}
+//	netConn, err := pool.Get()
+//	if err != nil {panic(err)}
+//	c := dns.Conn{Conn: netConn}
+//	q := dns.Msg{}
+//	q.SetQuestion("google.com.", dns.TypeA)
+//	log.Println(q)
+//	err = c.WriteMsg(&q)
+//	if err != nil {panic(err)}
+//	r, err := c.ReadMsg()
+//	if err != nil {panic(err)}
+//	log.Println(r)
+//	pool.Put(c.Conn)
 type TLSPool struct {
 	boot *bootstrapper
 
